Handle error of missing headers in file imported. Also, make header NOT case sensitive
