+++ core/templates/admin/csv_form.html
@@ -8,7 +8,7 @@
             {% csrf_token %}
 
             <button type="submit">Upload CSV</button>
-            <h2>The first row of csv file must have headers. The headers are case sensitive.</h2>
+            <h2>The first row of csv file must have headers.</h2>
             <p>The order of the columns is not important, but the following headers must exist:</p>
             <ul>
                 {{ form.header_list }}
